requirements.txt
opencv-python==4.10.0.84
backgroundremover==0.2.8
numpy==2.1.0
torchvision==0.19.0
torch==2.4.0
pillow==10.4.0